:recycle: 管理用エンドポイントを`/admin`以下へ分離
