+++ src/routes/admin.rs
@@ -1,2 +1,1 @@
-pub mod admin;
 pub mod plans;
